Refactored out FCGI into a separate parser
- Now have a parsers folder with both FCGI and HTTP parsers
