+++ SwiftCGI Framework/SwiftCGI/Server.swift
@@ -34,31 +34,35 @@
 // NOTE: This class muse inherit from NSObject; otherwise the Obj-C code for
 // GCDAsyncSocket will somehow not be able to store a reference to the delegate
 // (it will remain nil and no error will be logged).
-public class FCGIServer: NSObject, GCDAsyncSocketDelegate {
+public class FCGIServer: NSObject, GCDAsyncSocketDelegate, ParserDelegate {
     // MARK: Properties
     
     public let port: UInt16
-    public var paramsAvailableHandler: (Request -> Void)?
     public var requestRouter: Router
 //    public var requestHandler: FCGIRequestHandler
     
     let delegateQueue: dispatch_queue_t
-    var recordContext: [GCDAsyncSocket: FCGIRecord] = [:]
     private lazy var listener: GCDAsyncSocket = {
         GCDAsyncSocket(delegate: self, delegateQueue: self.delegateQueue)
     }()
     
-    private var currentRequests: [String: FCGIRequest] = [:]
+    
     private var activeSockets: Set<GCDAsyncSocket> = []
     
     private var registeredPreware: [RequestPrewareHandler] = []
     private var registeredMiddleware: [RequestMiddlewareHandler] = []
     private var registeredPostware: [RequestPostwareHandler] = []
-    private let httpParser = HTTPParser()
+    private let parser: Parser;
     
     // MARK: Init
     
     public init(port: UInt16, requestRouter: Router) {
+        if DEVSERVER {
+            parser = HTTPParser()
+        } else {
+            parser = FCGIParser()
+        }
+        
         self.port = port
         self.requestRouter = requestRouter
         
@@ -71,107 +75,6 @@ public class FCGIServer: NSObject, GCDAsyncSocketDelegate {
     public func start() throws {
         try listener.acceptOnPort(port)
     }
-    
-    func handleRecord(record: FCGIRecord, fromSocket socket: GCDAsyncSocket) {
-        let globalRequestID = "\(record.requestID)-\(socket.connectedPort)"
-        
-        // TODO: Guards to handle early exits in odd cases like malformed data; I don't like all of
-        // the forced unwrapping going on here right now...
-        
-        // Switch on record.type, since types can be mapped 1:1 to an FCGIRecord
-        // subclass. This allows for a much cleaner chunk of code than a handful
-        // of if/else ifs chained together, and allows the compiler to check that
-        // we have covered all cases
-        switch record.type {
-        case .BeginRequest:
-            let request = FCGIRequest(record: record as! BeginRequestRecord)
-            request.socket = socket
-            
-            objc_sync_enter(currentRequests)
-            currentRequests[globalRequestID] = request
-            objc_sync_exit(currentRequests)
-            
-            socket.readDataToLength(FCGIRecordHeaderLength, withTimeout: FCGITimeout, tag: FCGISocketTag.AwaitingHeaderTag.rawValue)
-        case .Params:
-            objc_sync_enter(currentRequests)
-            let maybeRequest = currentRequests[globalRequestID]
-            objc_sync_exit(currentRequests)
-            
-            if let request = maybeRequest {
-                if let params = (record as! ParamsRecord).params {
-                    if request._params == nil {
-                        request._params = [:]
-                    }
-                    
-                    // Copy the values into the request params dictionary
-                    for key in params.keys {
-                        request._params[key] = params[key]
-                    }
-                } else {
-                    paramsAvailableHandler?(request)
-                }
-                
-                socket.readDataToLength(FCGIRecordHeaderLength, withTimeout: FCGITimeout, tag: FCGISocketTag.AwaitingHeaderTag.rawValue)
-            }
-        case .Stdin:
-            objc_sync_enter(currentRequests)
-            let maybeRequest = currentRequests[globalRequestID]
-            objc_sync_exit(currentRequests)
-            
-            if var request = maybeRequest {
-                if request.streamData == nil {
-                    request.streamData = NSMutableData(capacity: 65536)
-                }
-                
-                if let recordData = (record as! ByteStreamRecord).rawData {
-                    request.streamData!.appendData(recordData)
-                } else {
-                    // TODO: Future - when Swift gets exception handling, wrap this
-                    // TODO: Refactor this into a separate method
-                    for handler in registeredPreware {
-                        request = handler(request) as! FCGIRequest  // Because we can't correctly force compiler type checking without generic typealiases
-                    }
-                    
-                    if let requestHandler = requestRouter.route(request.path) {
-                        if var response = requestHandler(request) {
-                            for handler in registeredMiddleware {
-                                response = handler(request, response)
-                            }
-                            
-                            if let responseData = response.responseData {
-                                request.writeResponseData(responseData, toStream: FCGIOutputStream.Stdout)
-                            }
-                            
-                            request.finish(.Complete)
-                            
-                            for handler in registeredPostware {
-                                handler(request, response)
-                            }
-                        } else {
-                            request.finish(.Complete)
-                            
-                            for handler in registeredPostware {
-                                handler(request, nil)
-                            }
-                        }
-                    }
-                    
-                    if let sock = request.socket {
-                        recordContext[sock] = nil
-                    }
-                    
-                    objc_sync_enter(currentRequests)
-                    currentRequests.removeValueForKey(globalRequestID)
-                    objc_sync_exit(currentRequests)
-                }
-            } else {
-                NSLog("WARNING: handleRecord called for invalid requestID")
-            }
-        default:
-            fatalError("ERROR: handleRecord called with an invalid FCGIRecord type")
-        }
-    }
-    
     
     // MARK: Pre/middle/postware registration
     
@@ -197,18 +100,12 @@ public class FCGIServer: NSObject, GCDAsyncSocketDelegate {
         let acceptedSocketQueue = dispatch_queue_create("SocketAcceptQueue-\(newSocket.connectedPort)", DISPATCH_QUEUE_SERIAL)
         newSocket.delegateQueue = acceptedSocketQueue
         
-        if DEVSERVER {
-            newSocket.readDataToData("\r\n".dataUsingEncoding(NSUTF8StringEncoding), withTimeout: 1000, tag: 0)
-        }
-        else {
-            newSocket.readDataToLength(FCGIRecordHeaderLength, withTimeout: FCGITimeout,
-                tag: FCGISocketTag.AwaitingHeaderTag.rawValue)
-        }
+        parser.resumeSocketReading(newSocket)
     }
     
     public func socketDidDisconnect(sock: GCDAsyncSocket?, withError err: NSError!) {
         if let sock = sock {
-            recordContext[sock] = nil
+            parser.socketDisconnect(sock)
             activeSockets.remove(sock)
         } else {
             NSLog("WARNING: nil sock disconnect")
@@ -216,42 +113,44 @@ public class FCGIServer: NSObject, GCDAsyncSocketDelegate {
     }
     
     public func socket(sock: GCDAsyncSocket!, didReadData data: NSData!, withTag tag: Int) {
-        if DEVSERVER {
-            print(NSString(data: data, encoding: NSUTF8StringEncoding)!)
-            httpParser.parseDataChunk(data)
-            sock.readDataWithTimeout(1000, tag: 0)
+        print(NSString(data: data, encoding: NSUTF8StringEncoding)!)
+        parser.parseData(sock, data: data, tag: tag)
+    }
+    
+    func finishedParsingRequest(request: Request) {
+        var req = request as! FCGIRequest
+        // TODO: Future - when Swift gets exception handling, wrap this
+        // TODO: Refactor this into a separate method
+        for handler in registeredPreware {
+            req = handler(request) as! FCGIRequest  // Because we can't correctly force compiler type checking without generic typealiases
         }
-        else {
-            if let socketTag = FCGISocketTag(rawValue: tag) {
-                switch socketTag {
-                case .AwaitingHeaderTag:
-                    // TODO: Clean up this
-                    // Phase 1 of 2 possible phases; first, try to parse the header
-                    if let record = createRecordFromHeaderData(data) {
-                        if record.contentLength == 0 {
-                            // No content; handle the message
-                            handleRecord(record, fromSocket: sock)
-                        } else {
-                            // Read additional content
-                            recordContext[sock] = record
-                            sock.readDataToLength(UInt(record.contentLength) + UInt(record.paddingLength), withTimeout: FCGITimeout, tag: FCGISocketTag.AwaitingContentAndPaddingTag.rawValue)
-                        }
-                    } else {
-                        NSLog("ERROR: Unable to construct request record")
-                        sock.disconnect()
-                    }
-                case .AwaitingContentAndPaddingTag:
-                    if let record = recordContext[sock] {
-                        record.processContentData(data)
-                        handleRecord(record, fromSocket: sock)
-                    } else {
-                        NSLog("ERROR: Case .AwaitingContentAndPaddingTag hit with no context")
-                    }
+        
+        if let requestHandler = requestRouter.route(req.path) {
+            if var response = requestHandler(req) {
+                for handler in registeredMiddleware {
+                    response = handler(req, response)
+                }
+                
+                if let responseData = response.responseData {
+                    req.writeResponseData(responseData, toStream: FCGIOutputStream.Stdout)
+                }
+                
+                req.finish(.Complete)
+                
+                for handler in registeredPostware {
+                    handler(req, response)
                 }
             } else {
-                NSLog("ERROR: Unknown socket tag")
-                sock.disconnect()
+                req.finish(.Complete)
+                
+                for handler in registeredPostware {
+                    handler(req, nil)
+                }
             }
+        }
+        
+        if let sock = request.socket {
+            parser.socketDisconnect(sock)
         }
     }
 }
